Simplify rendering logic.
Technique can reference to different texture views.
===================================================

In this way it is not neccessary anymore to have multiple
technique per back buffer in the ImageStreamRenderer.
One technique is enough and that one technique uses different image
views in its back buffer.

Remove rendering priority
=========================

It was a technical debt. Draw order recording doesn't makes any sense
because without barriers the commands will be executed in not
deterministic order.