Use Volk device table feature.
Loading vulkan prototype functions into the volk function table.

In this way multiple devices can be handled.

For ImGui it is more trickier. Imgui Uses its own function definitions
and these can be loaded via a function. The current solution is that
whenever the device is changes then the functions are reloaded. ImGui
anyways are only rendered when the given window is in focus.