C++23 and RenderTarget has Textures and TextureViews
C++23 is required for ranges::to

Textures can be play a wrapper roll from now. The goal is to be able to
track the textures resource state even if the texture is not owned by
the Texture class.

Current use case for the above mentioned situation when the texture is
owned by the swap chain.