WIP No more validation error.
Adding debugger for synchronization objects.

It is super useful to see which semaphore handle belongs to which
object. SyncObjects always have a proper purpose that can be describe.

Based on this the semaphores can be tracked and in case of validation
error it helps to understand what went wrong.

Decoupled the render_target_index from the synchronization objects.