Solve queue usage parallel issue.
We need to protect the queue againts parallel submit calls. A queue can
be used from multiple threads but not in the same time.

In case of imgui it requires extra care. Like if imgui would use the
queue to do something we need to protect it with the lock. Currently
imgui doesn't uses the queue.