WIP: Milestone reached: No validation error on release in DemoApp
The problem was around when only one queue is available. This queue
needs to be shared amongs the thread. But in this way this also needs to
be protected among the different use cases: transfer, render, present.
Because the same queue might be used.

Thus the load ballancer is shared amongs the command contexts.