WIP: SwapChain background worker
On profile it was clear that we have peek frames. On this frames the
sleep thread was sleeping much longer than 1ms.

The problem is around that we are fully dependent on the OS thread
handling.

The idea to avoid the peaks to organizing the calls into a background
thread.
The caller only places there a task and the result is a future. In this
way the power is more centralized and swap chain operations are not run
immedietly on the caller thread.
This makes only when there are more then one thread who makes render.

So in the future when there is only one rendering thread then it should
be a sync call on the thread.

When parallel frame rendering is enabled then it should use the
background thread.