Milestone: No more validation error
It works even when there are more threads then back buffer.

The fps in this case on the demo application is lower comparing when
num_of_threads<=back_buffer_count. It is due to the fact that we are
limited by the swap chain operations and having more threads that tries
to access to the swap chain just increases the cost of synchronization.