Fix transfer queue issue.
The problem was the following:
 When the transfer and render queue are not the same, ownership
 transforamation is needed. But, during the upload a wrong destination
 queue was used (destination was the transform queue. But it had no
 shader stage support, thus we have a validation error).

To solve the issue we needed a new SingleShotCommandBuffer factory for
the render queue. But it highlighted an old problem around the
CommandContext abstraction.
I.e.: CommandContext was not a context more a factory. Also it had
ownership issues. It showed with an enable_shared_from_this usage.

The solution is a new class CommandBufferContext. It owns the two
fatory. Also a VulkanQueue class is introduced. It is a new abstraction
level above the queue load ballancer.

Note: Current intel driver crashed when different presentation queue is
used then render queue.