WIP: Offscreen rendering.
RenderGraph builder is simplified. It was a bit too brute force. It
doesn't need to link againts all the following tasks. Also doesn't need
to link the predecessors. It was due to the idea of disabling nodes, but
it made too complex graph.

We can actually wait and signal the same semaphore during one submition.
For example timeline semaphores.

The pipeline looks like working, but better handle of upload/download is
necessary to avoid errors. (Read back reading back all the images that's
a problem. Need to separate the staging areas).